tests/wpt/web-platform-tests/payment-request/allowpaymentrequest/active-document-cross-origin.https.sub.html

<!doctype html>
<title>PaymentRequest &lt;iframe allowpaymentrequest> in non-active document (cross-origin)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/utils.js></script>
<iframe id="iframe" allowpaymentrequest></iframe>
<script>
async_test((t) => {
  const iframe = document.getElementById('iframe');
  const paymentArgs = [[{supportedMethods: 'foo'}], {total: {label: 'label', amount: {currency: 'USD', value: '5.00'}}}];

  onload = () => {
    const win = window[0];
    const grabbedPaymentRequest = win.PaymentRequest;
    win.location.href = make_absolute_url({
      hostname: '{{domains[www1]}}',
      path: '/common/blank.html'
    });
    iframe.onload = t.step_func_done(() => {
      assert_throws({name: 'SecurityError'}, () => {
        new grabbedPaymentRequest(...paymentArgs);
      });
    });
  }
});
</script>
